refactor: move aws signing bits to common/ (#9925)
This was under filter/http/common/aws for historical reasons, so moving
it to source/extensions/common/aws where it should live.
