+++ source/extensions/common/aws/BUILD
