+++ source/extensions/common/aws/region_provider_impl.cc
@@ -1,8 +1,7 @@
-#include "extensions/filters/http/common/aws/region_provider_impl.h"
+#include "extensions/common/aws/region_provider_impl.h"
 
 namespace Envoy {
 namespace Extensions {
-namespace HttpFilters {
 namespace Common {
 namespace Aws {
 
@@ -25,6 +24,5 @@ absl::optional<std::string> EnvironmentRegionProvider::getRegion() {
 
 } // namespace Aws
 } // namespace Common
-} // namespace HttpFilters
 } // namespace Extensions
 } // namespace Envoy
