+++ source/extensions/filters/http/aws_request_signing/BUILD
@@ -18,9 +18,9 @@ envoy_cc_library(
     hdrs = ["aws_request_signing_filter.h"],
     deps = [
         "//include/envoy/http:filter_interface",
+        "//source/extensions/common/aws:credentials_provider_impl_lib",
+        "//source/extensions/common/aws:signer_impl_lib",
         "//source/extensions/filters/http/common:pass_through_filter_lib",
-        "//source/extensions/filters/http/common/aws:credentials_provider_impl_lib",
-        "//source/extensions/filters/http/common/aws:signer_impl_lib",
         "@envoy_api//envoy/extensions/filters/http/aws_request_signing/v3:pkg_cc_proto",
     ],
 )
@@ -34,10 +34,10 @@ envoy_cc_extension(
     deps = [
         ":aws_request_signing_filter_lib",
         "//include/envoy/registry",
+        "//source/extensions/common/aws:credentials_provider_impl_lib",
+        "//source/extensions/common/aws:signer_impl_lib",
         "//source/extensions/filters/http:well_known_names",
         "//source/extensions/filters/http/common:factory_base_lib",
-        "//source/extensions/filters/http/common/aws:credentials_provider_impl_lib",
-        "//source/extensions/filters/http/common/aws:signer_impl_lib",
         "@envoy_api//envoy/extensions/filters/http/aws_request_signing/v3:pkg_cc_proto",
     ],
 )
