+++ source/extensions/filters/http/aws_request_signing/config.cc
@@ -4,10 +4,10 @@
 #include "envoy/extensions/filters/http/aws_request_signing/v3/aws_request_signing.pb.validate.h"
 #include "envoy/registry/registry.h"
 
+#include "extensions/common/aws/credentials_provider_impl.h"
+#include "extensions/common/aws/signer_impl.h"
+#include "extensions/common/aws/utility.h"
 #include "extensions/filters/http/aws_request_signing/aws_request_signing_filter.h"
-#include "extensions/filters/http/common/aws/credentials_provider_impl.h"
-#include "extensions/filters/http/common/aws/signer_impl.h"
-#include "extensions/filters/http/common/aws/utility.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -19,9 +19,9 @@ Http::FilterFactoryCb AwsRequestSigningFilterFactory::createFilterFactoryFromPro
     const std::string& stats_prefix, Server::Configuration::FactoryContext& context) {
 
   auto credentials_provider =
-      std::make_shared<HttpFilters::Common::Aws::DefaultCredentialsProviderChain>(
-          context.api(), HttpFilters::Common::Aws::Utility::metadataFetcher);
-  auto signer = std::make_unique<HttpFilters::Common::Aws::SignerImpl>(
+      std::make_shared<Extensions::Common::Aws::DefaultCredentialsProviderChain>(
+          context.api(), Extensions::Common::Aws::Utility::metadataFetcher);
+  auto signer = std::make_unique<Extensions::Common::Aws::SignerImpl>(
       config.service_name(), config.region(), credentials_provider,
       context.dispatcher().timeSource());
