+++ source/extensions/grpc_credentials/aws_iam/BUILD
@@ -25,10 +25,10 @@ envoy_cc_extension(
         "//source/common/grpc:google_grpc_creds_lib",
         "//source/common/http:message_lib",
         "//source/common/http:utility_lib",
-        "//source/extensions/filters/http/common/aws:credentials_provider_impl_lib",
-        "//source/extensions/filters/http/common/aws:region_provider_impl_lib",
-        "//source/extensions/filters/http/common/aws:signer_impl_lib",
-        "//source/extensions/filters/http/common/aws:utility_lib",
+        "//source/extensions/common/aws:credentials_provider_impl_lib",
+        "//source/extensions/common/aws:region_provider_impl_lib",
+        "//source/extensions/common/aws:signer_impl_lib",
+        "//source/extensions/common/aws:utility_lib",
         "//source/extensions/grpc_credentials:well_known_names",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/grpc_credential/v3:pkg_cc_proto",
