+++ test/extensions/common/aws/BUILD
@@ -14,8 +14,8 @@ envoy_cc_mock(
     srcs = ["mocks.cc"],
     hdrs = ["mocks.h"],
     deps = [
-        "//source/extensions/filters/http/common/aws:credentials_provider_interface",
-        "//source/extensions/filters/http/common/aws:signer_interface",
+        "//source/extensions/common/aws:credentials_provider_interface",
+        "//source/extensions/common/aws:signer_interface",
     ],
 )
 
@@ -25,8 +25,8 @@ envoy_cc_test(
     deps = [
         "//source/common/buffer:buffer_lib",
         "//source/common/http:message_lib",
-        "//source/extensions/filters/http/common/aws:signer_impl_lib",
-        "//test/extensions/filters/http/common/aws:aws_mocks",
+        "//source/extensions/common/aws:signer_impl_lib",
+        "//test/extensions/common/aws:aws_mocks",
         "//test/test_common:simulated_time_system_lib",
         "//test/test_common:utility_lib",
     ],
@@ -36,7 +36,7 @@ envoy_cc_test(
     name = "utility_test",
     srcs = ["utility_test.cc"],
     deps = [
-        "//source/extensions/filters/http/common/aws:utility_lib",
+        "//source/extensions/common/aws:utility_lib",
         "//test/test_common:utility_lib",
     ],
 )
@@ -45,7 +45,7 @@ envoy_cc_test(
     name = "region_provider_impl_test",
     srcs = ["region_provider_impl_test.cc"],
     deps = [
-        "//source/extensions/filters/http/common/aws:region_provider_impl_lib",
+        "//source/extensions/common/aws:region_provider_impl_lib",
         "//test/test_common:environment_lib",
     ],
 )
@@ -54,8 +54,8 @@ envoy_cc_test(
     name = "credentials_provider_impl_test",
     srcs = ["credentials_provider_impl_test.cc"],
     deps = [
-        "//source/extensions/filters/http/common/aws:credentials_provider_impl_lib",
-        "//test/extensions/filters/http/common/aws:aws_mocks",
+        "//source/extensions/common/aws:credentials_provider_impl_lib",
+        "//test/extensions/common/aws:aws_mocks",
         "//test/mocks/api:api_mocks",
         "//test/mocks/event:event_mocks",
         "//test/test_common:environment_lib",
@@ -67,7 +67,7 @@ envoy_cc_test(
     name = "credentials_provider_test",
     srcs = ["credentials_provider_test.cc"],
     deps = [
-        "//source/extensions/filters/http/common/aws:credentials_provider_interface",
+        "//source/extensions/common/aws:credentials_provider_interface",
     ],
 )
 
@@ -78,7 +78,7 @@ envoy_cc_test(
     ],
     deps = [
         "//source/common/common:fmt_lib",
-        "//source/extensions/filters/http/common/aws:utility_lib",
+        "//source/extensions/common/aws:utility_lib",
         "//source/extensions/filters/http/fault:config",
         "//source/extensions/filters/http/fault:fault_filter_lib",
         "//source/extensions/filters/http/router:config",
