+++ test/extensions/common/aws/credentials_provider_impl_test.cc
@@ -1,6 +1,6 @@
-#include "extensions/filters/http/common/aws/credentials_provider_impl.h"
+#include "extensions/common/aws/credentials_provider_impl.h"
 
-#include "test/extensions/filters/http/common/aws/mocks.h"
+#include "test/extensions/common/aws/mocks.h"
 #include "test/mocks/api/mocks.h"
 #include "test/mocks/event/mocks.h"
 #include "test/test_common/environment.h"
@@ -14,7 +14,6 @@ using testing::Return;
 
 namespace Envoy {
 namespace Extensions {
-namespace HttpFilters {
 namespace Common {
 namespace Aws {
 
@@ -444,6 +443,5 @@ TEST(CredentialsProviderChainTest, getCredentials_secondProviderReturns) {
 
 } // namespace Aws
 } // namespace Common
-} // namespace HttpFilters
 } // namespace Extensions
 } // namespace Envoy
