+++ test/extensions/common/aws/signer_impl_test.cc
@@ -1,10 +1,10 @@
 #include "common/buffer/buffer_impl.h"
 #include "common/http/message_impl.h"
 
-#include "extensions/filters/http/common/aws/signer_impl.h"
-#include "extensions/filters/http/common/aws/utility.h"
+#include "extensions/common/aws/signer_impl.h"
+#include "extensions/common/aws/utility.h"
 
-#include "test/extensions/filters/http/common/aws/mocks.h"
+#include "test/extensions/common/aws/mocks.h"
 #include "test/test_common/simulated_time_system.h"
 #include "test/test_common/utility.h"
 
@@ -13,7 +13,6 @@ using testing::Return;
 
 namespace Envoy {
 namespace Extensions {
-namespace HttpFilters {
 namespace Common {
 namespace Aws {
 namespace {
@@ -167,6 +166,5 @@ TEST_F(SignerImplTest, SignHostHeader) {
 } // namespace
 } // namespace Aws
 } // namespace Common
-} // namespace HttpFilters
 } // namespace Extensions
 } // namespace Envoy
